OSRS Animation Configurations
Added a int array to be able to easily grab animation definitions from
the OSRS cache. Inside Animations.java int[] osrsAnimations. Still
requires the actual animation files to be added in order for the OSRS
animations to work.

The process of adding OSRS Animations:
Figure out ur animation id
Add the animation id to the int[] array
Add the animation id to the debug int[] array
Grab the Sys.out for the file id
Pack the animation in the cache
Start your client

@JonathanSirens